requirements.txt
chris_plugin==0.1.2
pydicom
Pillow
numpy
loguru
pftag==1.2.22
pflog==1.2.26
pftel-client
